adapter: Create deterministic log index IDs
This commit adds a new variant of `GlobalId` and `CatalogItemId` for
introspection source indexes. The values of these IDs are
deterministically derived from the cluster ID and the log variant.

Introspection source indexes are a special edge case of items. They are
considered system items, but they are the only system item that can be
created by the user at any time. All other system items can only be
created by the system during the startup of an upgrade.

Previously, it was possible to allocate the same System ID to two
different objects if something like the following happened:

1. Materialize version `v` is running in read-write mode.
2. Materialize version `v + 1` starts in read-only mode.
3. The next system item ID is `s`.
4. `v + 1` allocates `s` for a new system item (table, view,
   introspection source, etc.)
5. `v` creates a new user cluster and allocates `s` through `s + n` to
   the introspection source indexes in that cluster. At this point we
   have two separate objects with the same Global ID, which is bad.
6. `v + 1` reboots in read-write mode and allocates `s + n + 1` to the
   new system item. At this point the new system item has received two
   different IDs, which is also bad.

Putting introspection source index IDs in their own namespace and
making them deterministic removes this issue and ones like it.

Fixes #MaterializeInc/database-issues/issues/8731

+++ src/storage-types/src/instances.rs
@@ -25,13 +25,13 @@ include!(concat!(env!("OUT_DIR"), "/mz_storage_types.instances.rs"));
 )]
 pub enum StorageInstanceId {
     /// A system storage instance.
-    System(u64),
+    System(u32),
     /// A user storage instance.
-    User(u64),
+    User(u32),
 }
 
 impl StorageInstanceId {
-    pub fn inner_id(&self) -> u64 {
+    pub fn inner_id(&self) -> u32 {
         match self {
             StorageInstanceId::System(id) | StorageInstanceId::User(id) => *id,
         }
@@ -53,7 +53,7 @@ impl FromStr for StorageInstanceId {
         if s.len() < 2 {
             bail!("couldn't parse compute instance id {}", s);
         }
-        let val: u64 = s[1..].parse()?;
+        let val: u32 = s[1..].parse()?;
         match s.chars().next().unwrap() {
             's' => Ok(Self::System(val)),
             'u' => Ok(Self::User(val)),
